refactor: reorganize session-related schemas and utilities for better structure
