+++ apps/server/src/utils/index.ts
@@ -1,3 +1,4 @@
 export * from './constants';
 export * from './converters';
 export * from './cookie';
+export * from './session';
