+++ packages/shared/src/models/index.ts
@@ -1,6 +1,8 @@
+export * from './auth';
 export * from './errors';
 export * from './file';
 export * from './folder';
 export * from './generalPreference';
 export * from './notification';
+export * from './session';
 export * from './user';
