+++ packages/shared/src/models/user.ts
@@ -29,35 +29,4 @@ export const userSchema = z.object({
 
 export type User = z.infer<typeof userSchema>;
 
-export const userSessionSchema = z.object({
-  id: z.number().int().nonnegative(),
-  userId: z.number().int().nonnegative(),
-  deviceInfo: z.string().min(1),
-  ip: z.string().ip(),
-  cookie: z.string().nullable(),
-  createdAt: z.date(),
-  lastActive: z.date(),
-  loginAttempts: z.number().int().nonnegative(),
-  banDurationMinutes: z.number().int().nullable(), // null = no ban, -1 = permanent ban
-  banStart: z.date().nullable()
-});
 
-export type UserSession = z.infer<typeof userSessionSchema>;
-
-export const userSessionCreateSchema = userSessionSchema.omit({ id: true });
-
-export type UserSessionCreate = z.infer<typeof userSessionCreateSchema>;
-
-export const userRegisterSchema = userSchema.omit({ id: true, role: true, username: true, createdAt: true });
-
-export type UserRegister = z.infer<typeof userRegisterSchema>;
-
-export const authResponseSchema = userSchema.omit({ id: true, password: true, createdAt: true });
-
-export type AuthResponse = z.infer<typeof authResponseSchema>;
-
-export const userLoginSchema = userSchema.pick({
-  email: true,
-  password: true
-});
-export type UserLogin = z.infer<typeof userLoginSchema>;
